# api/swagger/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "gateway.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/gateways/{gateway_id}/frames": {
      "get": {
        "summary": "StreamFrameLogs streams the uplink and downlink frame-logs for the given mac.\nNote: these are the raw LoRaWAN frames and this endpoint is intended for debugging.",
        "operationId": "StreamFrameLogs",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/apiStreamGatewayFrameLogsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "gateway_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "GatewayService"
        ]
      }
    },
...
    "/api/gateways/{gateway_id}/stats": {
      "get": {
        "summary": "GetStats lists the gateway stats given the query parameters.",
        "operationId": "GetStats",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiGetGatewayStatsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "gateway_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "interval",
            "description": "Aggregation interval.  One of \"second\", \"minute\", \"hour\", \"day\", \"week\",\n\"month\", \"quarter\", \"year\".  Case insensitive.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "start_timestamp",
            "description": "Timestamp to start from.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "end_timestamp",
            "description": "Timestamp until to get from.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "tags": [
          "GatewayService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiDownlinkFrameLog": {
      "type": "object",
      "properties": {
        "txInfo": {
          "$ref": "#/definitions/apiDownlinkTXInfo",
          "description": "TX information of the downlink."
        },
        "phyPayloadJSON": {
          "type": "string",
          "description": "LoRaWAN PHYPayload."
        }
      }
    },
    "apiDownlinkTXInfo": {
      "type": "object",
      "properties": {
        "gatewayId": {
          "type": "string",
          "description": "Gateway ID."
        },
        "immediately": {
          "type": "boolean",
          "format": "boolean",
          "description": "Frame must be sent immediately."
        },
        "timeSinceGpsEpoch": {
          "type": "string",
          "description": "Emit frame at the given time since GPS epoch."
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "description": "Emit the frame at the given gateway internal timestamp."
        },
        "frequency": {
          "type": "integer",
          "format": "int64",
          "description": "TX frequency (in Hz)."
        },
        "power": {
          "type": "integer",
          "format": "int32",
          "description": "TX power (in dBm)."
        },
        "modulation": {
          "$ref": "#/definitions/commonModulation",
          "description": "Modulation."
        },
        "loraModulationInfo": {
          "$ref": "#/definitions/gwLoRaModulationInfo",
          "description": "LoRa modulation information."
        },
        "fskModulationInfo": {
          "$ref": "#/definitions/gwFSKModulationInfo",
          "description": "FSK modulation information."
        },
        "board": {
          "type": "integer",
          "format": "int64",
          "description": "The board identifier for emitting the frame."
        },
        "antenna": {
          "type": "integer",
          "format": "int64",
          "description": "The antenna identifier for emitting the frame."
        }
      },
      "description": "Same comment as above applies to this message."
    },
...
    "apiGatewayStats": {
      "type": "object",
      "properties": {
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp of the (aggregated) measurement."
        },
        "rxPacketsReceived": {
          "type": "integer",
          "format": "int32",
          "description": "Packets received by the gateway."
        },
        "rxPacketsReceivedOK": {
          "type": "integer",
          "format": "int32",
          "description": "Packets received by the gateway that passed the CRC check."
        },
        "txPacketsReceived": {
          "type": "integer",
          "format": "int32",
          "description": "Packets received by the gateway for transmission."
        },
        "txPacketsEmitted": {
          "type": "integer",
          "format": "int32",
          "description": "Packets transmitted by the gateway."
        }
      }
    },
...
    "apiGetGatewayStatsResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiGatewayStats"
          }
        }
      }
    },
...
    "apiStreamGatewayFrameLogsResponse": {
      "type": "object",
      "properties": {
        "uplinkFrame": {
          "$ref": "#/definitions/apiUplinkFrameLog",
          "description": "Contains an uplink frame."
        },
        "downlinkFrame": {
          "$ref": "#/definitions/apiDownlinkFrameLog",
          "description": "Contains a downlink frame."
        }
      }
    },
...
    "apiUplinkFrameLog": {
      "type": "object",
      "properties": {
        "txInfo": {
          "$ref": "#/definitions/gwUplinkTXInfo",
          "description": "TX information of the uplink."
        },
        "rxInfo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUplinkRXInfo"
          },
          "description": "RX information of the uplink."
        },
        "phyPayloadJSON": {
          "type": "string",
          "description": "LoRaWAN PHYPayload."
        }
      }
    },
    "apiUplinkRXInfo": {
      "type": "object",
      "properties": {
        "gatewayId": {
          "type": "string",
          "description": "Gateway ID."
        },
        "time": {
          "type": "string",
          "format": "date-time",
          "description": "RX time (only set when the gateway has a GPS module)."
        },
        "timeSinceGpsEpoch": {
          "type": "string",
          "description": "RX time since GPS epoch (only set when the gateway has a GPS module)."
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "description": "Gateway internal timestamp."
        },
        "rssi": {
          "type": "integer",
          "format": "int32",
          "description": "RSSI."
        },
        "loraSnr": {
          "type": "number",
          "format": "double",
          "description": "LoRa SNR."
        },
        "channel": {
          "type": "integer",
          "format": "int64",
          "description": "Channel."
        },
        "rfChain": {
          "type": "integer",
          "format": "int64",
          "description": "RF Chain."
        },
        "board": {
          "type": "integer",
          "format": "int64",
          "description": "Board."
        },
        "antenna": {
          "type": "integer",
          "format": "int64",
          "description": "Antenna."
        },
        "location": {
          "$ref": "#/definitions/gwLocation",
          "description": "Location."
        }
      },
      "description": "This is a copy of gw.UplinkRXInfo with the only change that the\ngateway_id is of type string so that we can return it as HEX encoded\ninstead of base64."
    },
    "commonModulation": {
      "type": "string",
      "enum": [
        "LORA",
        "FSK"
      ],
      "default": "LORA",
      "title": "- LORA: LoRa\n - FSK: FSK"
    },
    "gwFSKModulationInfo": {
      "type": "object",
      "properties": {
        "bandwidth": {
          "type": "integer",
          "format": "int64",
          "description": "Bandwidth."
        },
        "bitrate": {
          "type": "integer",
          "format": "int64",
          "description": "Bitrate."
        }
      }
    },
    "gwLoRaModulationInfo": {
      "type": "object",
      "properties": {
        "bandwidth": {
          "type": "integer",
          "format": "int64",
          "description": "Bandwidth."
        },
        "spreadingFactor": {
          "type": "integer",
          "format": "int64",
          "description": "Speading-factor."
        },
        "codeRate": {
          "type": "string",
          "description": "Code-rate."
        },
        "polarizationInversion": {
          "type": "boolean",
          "format": "boolean",
          "description": "Polarization inversion."
        }
      }
    },
    "gwLocation": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double",
          "description": "Latitude."
        },
        "longitude": {
          "type": "number",
          "format": "double",
          "description": "Longitude."
        },
        "altitude": {
          "type": "number",
          "format": "double",
          "description": "Altitude."
        }
      }
    },
    "gwUplinkTXInfo": {
      "type": "object",
      "properties": {
        "frequency": {
          "type": "integer",
          "format": "int64",
          "description": "Frequency (Hz)."
        },
        "modulation": {
          "$ref": "#/definitions/commonModulation",
          "description": "Modulation."
        },
        "loraModulationInfo": {
          "$ref": "#/definitions/gwLoRaModulationInfo",
          "description": "LoRa modulation information."
        },
        "fskModulationInfo": {
          "$ref": "#/definitions/gwFSKModulationInfo",
          "description": "FSK modulation information."
        }
      }
    },
...
  }
}